feat(qgrads): moved to plugin
breaking change
