+++ docs/creator/plugins/index.md
@@ -5,4 +5,5 @@ The Elastic-AI.creator comes bundled with a few plugins.
 
 ```{toctree}
 vhdl
+python
 ```
