+++ docs/creator/plugins/python.md
@@ -1,0 +1,5 @@
+# Python Plugins
+
+```{toctree}
+quantized_grads
+```
