+++ docs/creator/plugins/quantized_grads.md
@@ -1,0 +1,2 @@
+```{include} ../../../elasticai/creator_plugins/quantized_grads/README.md
+```
