+++ elasticai/creator_plugins/quantized_grads/meta.toml
@@ -1,0 +1,8 @@
+
+[[plugins]]
+name = 'quantized_grads'
+version = '0.1'
+api_version = '0.1'
+target_runtime = 'python'
+target_platform = 'native'
+
