+++ elasticai/creator_plugins/quantized_grads/tests/fxpmodel_loss_optim_test.py
@@ -3,13 +3,15 @@ import torch
 from torch.nn import Sequential
 from torch.optim import SGD
 
-from elasticai.creator.nn.quantized_grads.base_modules import Linear
-from elasticai.creator.nn.quantized_grads.fixed_point import (
+from elasticai.creator_plugins.quantized_grads.base_modules import Linear
+from elasticai.creator_plugins.quantized_grads.fixed_point import (
     FixedPointConfigV2,
     QuantizeForwHTEBackwHTE,
     QuantizeParamToFixedPointHTE,
 )
-from elasticai.creator.nn.quantized_grads.quantized_optim import get_quantized_optimizer
+from elasticai.creator_plugins.quantized_grads.quantized_optim import (
+    get_quantized_optimizer,
+)
 
 
 class Test1:
