+++ elasticai/creator_plugins/quantized_grads/tests/test_linear.py
@@ -1,5 +1,7 @@
-from elasticai.creator.nn.quantized_grads.base_modules import Conv1d
-from elasticai.creator.nn.quantized_grads.fixed_point import (
+import torch
+
+from elasticai.creator_plugins.quantized_grads.base_modules import Linear
+from elasticai.creator_plugins.quantized_grads.fixed_point import (
     FixedPointConfigV2,
     QuantizeForwHTE,
     QuantizeParamToFixedPointHTE,
@@ -10,11 +12,15 @@ from elasticai.creator.nn.quantized_grads.fixed_point import (
 def test_conv1d_fxp_init():
     conf = FixedPointConfigV2(8, 3)
 
-    Conv1d(
+    l = Linear(
         QuantizeForwHTE(conf),
-        QuantizeParamToFixedPointHTE(conf),
+        2,
         3,
-        2,
-        2,
+        QuantizeParamToFixedPointHTE(conf),
+        True,
         bias_quantization=QuantizeParamToFixedPointStochastic(conf),
     )
+
+    # print(l.)
+    x = torch.randn(1, 1, 2)
+    l(x)
