+++ elasticai/creator_plugins/quantized_grads/tests/two_complement_fixed_point_config_test.py
@@ -2,7 +2,7 @@ import dataclasses
 
 import pytest
 
-from elasticai.creator.nn.quantized_grads.fixed_point.two_complement_fixed_point_config import (
+from elasticai.creator_plugins.quantized_grads.fixed_point import (
     FixedPointConfigV2,
 )
